test(#13) Reestruturando testes
